Refactor SessionStorage helper functions to use ICartItem
This commit refactors the saveCartInSessionStorage and getCartFromSessionStorage helper functions. These functions now use just the ICartItem interface, instead of a custom Cart interface. This reduces complexity and increases reusability of the code. It also aligns better with the use of these functions mostly dealing with individual cart items rather than the whole cart.