Add product count and subtotal to CartTable
An additional 'count' column was added to the CartTable component to show the quantity of each product in the cart. Similarly, a 'subtotal' column was also added to show the total cost per product type. This required changing IProduct to IProductWithCount in the map function to properly render count and subtotal. The 'Total del Pedido' display is now right-aligned.