"Update IProductTable interface to be more flexible"
Updated the IProductTable interface to support a dynamic context type and allow products' variable to be passed as a prop. This change reduces redundancy and increases flexibility by ensuring that the table can be used within different contexts (admin or cart) and display varied products based on the requirements of the parent component.