Add check and component for empty shopping cart
In order to improve user experience, a new EmptyCart component has been created and integrated into the ProductsTableContainer. This component comes into play when there are no items in the user's shopping cart. It displays a user-friendly message prompting the user to continue shopping, a link to the homepage and a placeholder image.