Add Google CDN to Next.js config and update user avatar path
This commit includes the addition of Google's CDN to the Next.js configuration for an additional static file host. It also corrects the default path to the user avatar in the Header component within the admin panel. These modifications improve the application's performance and fix the issue with the default user avatar not loading correctly.