Fix user avatar path and image loading issue in Header component
In this commit, the path to the user avatar in the Header component was corrected and the issue regarding the avatar not loading correctly was fixed. This was achieved by modifying the static file host in Next.js configuration by adding Google's CDN. These changes not only correct image path errors but also help to improve the overall application's performance.